main: show directory contents
It's nice to have more visual confirmation that an upload was
successful, so I've added the directory contents of the directory that
uploads are being streamed to on the index page. It also shows the file
sizes, since that also helps with confirmation.

The next step will be to optionally allow downloading of files from the
directory, if this wants to be used in the opposite direction, like
Python's SimpleHTTPServer.